+++ quake3-rs/uix86_64/src/q3_ui/ui_demo2.rs
@@ -759,7 +759,7 @@ unsafe extern "C" fn Demos_MenuInit() {
                     (::std::mem::size_of::<[libc::c_char; 32768]>() as libc::c_ulong)
                         .wrapping_div(::std::mem::size_of::<libc::c_char>() as libc::c_ulong)
                         .wrapping_sub(demoname.offset_from(s_demos.names.as_mut_ptr())
-                            as libc::c_long as libc::c_ulong) as i32,
+                            as isize as libc::c_ulong) as i32,
                 )
         }
         j += 1
